feat: standardized-provider-neutral-conductor-007 - Prove concurrent cancellation, dispatch, and writer backpressure
